добавил начисление, получение и историю для клиник

+++ src/main/java/com/alpengotter/dental_service_bank/domain/dto/ClinicCurrencyUpdateDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class ClinicCurrencyUpdateDto {
-    private Integer currency;
+    private Long currency;
     private String comment;
 }
